Added ElementUtils class for Elements names increase flexibility.
Changed the metamodel definition a little bit. Added type for control
flow, lifelines and messages too. Maybe a good idea is to make this a
standard attribute for every element.

Added ElementNames enumeration. It seems a good option for future
growth.

Won't make anymore changes for a while. Still needs javadoc coments and
non-abuse of static methods like it is now.